print fixes & node type changes

+++ Node.h
@@ -1,10 +1,11 @@
 #pragma once
 
+
 class Node
 {
 public:
-	int from;
-	int d;
+	short int from;
+	short int d;
 	Node();
 	Node(int d, int from);
 };
